# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cohort — AI agent fleet governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{agentId}/verdict/decision:
    post:
      operationId: decideVerdict
      tags: [agents]
      summary: Approve, disagree with, or export the current verdict
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerdictDecisionInput"
      responses:
        "200":
          description: Updated verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verdict"
...
  /discovery/pre-assess:
    post:
      operationId: preAssessAgentSource
      tags: [agents]
      summary: Heuristic pre-assessment of an agent repository (no AI key needed)
      description: >-
        Fetches the repository, understands the agent via static heuristics
        (stack detection, README dossier parsing, KPI-table extraction) and
        returns a pre-filled Carteira de Trabalho draft with per-field
        confidence, framing missing metrics from the metric catalog. Works
        without any AI credential; when one is configured the AI analyzer can
        refine this draft afterwards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreAssessInput"
      responses:
        "200":
          description: Pre-assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreAssessResult"
...
  /agents/{agentId}/events:
    post:
      operationId: ingestAgentEvent
      tags: [agents]
      summary: Ingest a telemetry event (execution, error, escalation) for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentEventInput"
      responses:
        "202":
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted:
                    type: boolean
  /integrations/agent-events:
    post:
      operationId: ingestExternalAgentEnvelope
      tags: [agents]
      summary: Ingest a normalized external agent envelope
      description: >-
        Universal ingestion contract for platforms such as Zendesk,
        Agentforce, OpenTelemetry gateways and proprietary runtimes. The
        envelope separates identity, execution telemetry, metric evidence and
        human feedback so the source adapter can evolve without changing the
        Muster evaluation model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalAgentEnvelope"
      responses:
        "202":
          description: External envelope accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalAgentIngestionResult"
  /agents/{agentId}/heartbeat:
    post:
      operationId: reportAgentHeartbeat
      tags: [agents]
      summary: Report that an agent runtime is alive and supervised
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentHeartbeatInput"
      responses:
        "202":
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentHeartbeatAccepted"
  /agents/{agentId}/supervision:
    get:
      operationId: readAgentSupervision
      tags: [agents]
      summary: Read live supervision freshness for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Supervision freshness
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSupervision"
  /agents/{agentId}/telemetry/{window}:
    get:
      operationId: getAgentTelemetry
      tags: [agents]
      summary: Aggregated telemetry summary for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: window
          in: path
          required: true
          schema:
            type: string
            enum: ["7d", "30d", "90d"]
      responses:
        "200":
          description: Telemetry summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetrySummary"
  /agents/{agentId}/reevaluate:
    post:
      operationId: reevaluateAgent
      tags: [agents]
      summary: Recompute the agent evaluation — from real telemetry when available
      responses:
        "200":
          description: Reevaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReevaluateOutcome"
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
components:
  schemas:
...
    NextAction:
      type: object
      required:
        - action
        - owner
        - due
      properties:
        action:
          type: string
        owner:
          type: string
        due:
          type: string
    Verdict:
      type: object
      required:
        - id
        - agentId
        - verdict
        - confidence
        - executionWindow
        - suggestedSponsor
        - nextActions
        - rationale
        - decision
        - createdAt
      properties:
        id:
          type: string
        agentId:
          type: string
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        confidence:
          type: number
        executionWindow:
          type: string
        suggestedSponsor:
          type: string
        nextActions:
          type: array
          items:
            $ref: "#/components/schemas/NextAction"
        rationale:
          type: string
        decision:
          type: string
          enum: [pending, approved, disagreed, exported]
        decidedBy:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ExternalAgentEnvelope:
      type: object
      required: [contractVersion, source, agent]
      properties:
        contractVersion:
          type: string
          enum: [muster.agent-ingestion.v1]
        eventId:
          type: string
          description: Optional idempotency key supplied by the source platform.
        source:
          type: object
          required: [platform]
          properties:
            platform:
              type: string
            tenant:
              type: string
            connectorId:
              type: string
            environment:
              type: string
            reference:
              type: string
        agent:
          type: object
          required: [externalId, name]
          properties:
            externalId:
              type: string
            name:
              type: string
            version:
              type: string
            runtime:
              type: string
            url:
              type: string
            metadata:
              type: object
              additionalProperties: true
        execution:
          $ref: "#/components/schemas/ExternalAgentExecution"
        observations:
          type: array
          items:
            $ref: "#/components/schemas/ExternalAgentObservation"
        feedback:
          $ref: "#/components/schemas/ExternalAgentFeedback"
    ExternalAgentExecution:
      type: object
      properties:
        id:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [success, error, escalated, running]
        durationMs:
          type: number
        costCents:
          type: number
        tokensIn:
          type: number
        tokensOut:
          type: number
        metadata:
          type: object
          additionalProperties: true
    ExternalAgentObservation:
      type: object
      required: [metricKey, label, value, unit]
      properties:
        metricKey:
          type: string
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        kind:
          type: string
          enum: [observed, inferred, synthetic]
        confidence:
          type: number
        sampleSize:
          type: integer
        capturedAt:
          type: string
          format: date-time
        lineage:
          type: array
          items:
            type: object
            required: [stage, name]
            properties:
              stage:
                type: string
              name:
                type: string
              ref:
                type: string
    ExternalAgentFeedback:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [positive, negative, correction, escalation]
        score:
          type: number
        comment:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ExternalAgentIngestionResult:
      type: object
      required: [accepted, mapped, agentExternalId, eventsAccepted, observationsAccepted, nextAction]
      properties:
        accepted:
          type: boolean
        mapped:
          type: boolean
        agentExternalId:
          type: string
        agentId:
          type: [string, "null"]
        eventsAccepted:
          type: integer
        observationsAccepted:
          type: integer
        nextAction:
          type: string
...
    VerdictDecisionInput:
      type: object
      required:
        - decision
      properties:
        decision:
          type: string
          enum: [approved, disagreed, exported]
        reason:
          type: string
        decidedBy:
          type: string
    DraftMetric:
      type: object
      required:
        - layer
        - label
        - unit
        - target
      properties:
        layer:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        unit:
          type: string
        target:
          type: string
        value:
          type: number
          description: >-
            Optional reviewer-set starting/current value for the metric. When
            provided during admission it overrides the deterministically seeded
            value so goal-vs-actual reflects reality.
        rationale:
          type: string
    DraftBusinessCase:
      type: object
      required:
        - baseline
        - targetPayback
        - description
      properties:
        baseline:
          type: string
        targetPayback:
          type: string
        description:
          type: string
...
    AgentDraft:
      type: object
      required:
        - name
        - role
        - tagline
        - bio
        - shouldDo
        - shouldNotDo
        - autonomyLevel
        - limits
        - businessCase
        - proposedMetrics
        - summary
        - confidence
      properties:
        name:
          type: string
        role:
          type: string
        tagline:
          type: string
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/DraftBusinessCase"
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/DraftMetric"
        summary:
          type: string
        confidence:
          type: number
...
    PreAssessInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 8
        nameHint:
          type: string
    PreAssessResult:
      type: object
      required: [draft, fieldConfidence, platform, signals]
      properties:
        draft:
          $ref: "#/components/schemas/AgentDraft"
        fieldConfidence:
          type: object
          additionalProperties:
            type: number
        platform:
          type: ["string", "null"]
        signals:
          type: array
          items:
            type: string
...
    AgentEventInput:
      type: object
      properties:
        kind:
          type: string
          enum: [execution, error, escalation, feedback, heartbeat]
        # Optional occurrence time (ISO 8601). Agents that batch or replay
        # events report the real execution time; when omitted the server
        # stamps the event with "now".
        ts:
          type: string
          format: date-time
        durationMs:
          type: number
        costCents:
          type: number
        tokensIn:
          type: number
        tokensOut:
          type: number
        success:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    AgentHeartbeatInput:
      type: object
      properties:
        runtime:
          type: string
        version:
          type: string
        intervalSeconds:
          type: number
        status:
          type: string
          enum: [healthy, degraded, stopped]
        metadata:
          type: object
          additionalProperties: true
    AgentHeartbeatAccepted:
      type: object
      required: [accepted, observedAt]
      properties:
        accepted:
          type: boolean
        observedAt:
          type: string
          format: date-time
    AgentSupervision:
      type: object
      required: [agentId, status, isStale, intervalSeconds, ageSeconds, lastHeartbeatAt]
      properties:
        agentId:
          type: string
        status:
          type: string
          enum: [live, delayed, stale, unknown]
        isStale:
          type: boolean
        intervalSeconds:
          type: number
        ageSeconds:
          type: number
        lastHeartbeatAt:
          type: [string, "null"]
        runtime:
          type: [string, "null"]
        version:
          type: [string, "null"]
        reportedStatus:
          type: [string, "null"]
    TelemetrySummary:
      type: object
      required:
        - windowDays
        - totalExecutions
        - totalCostCents
        - executionsPerDay
        - activeDays
      properties:
        windowDays:
          type: number
        totalExecutions:
          type: number
        successRate:
          type: ["number", "null"]
        avgDurationMs:
          type: ["number", "null"]
        p95DurationMs:
          type: ["number", "null"]
        totalCostCents:
          type: number
        avgCostCentsPerExecution:
          type: ["number", "null"]
        executionsPerDay:
          type: number
        escalationRate:
          type: ["number", "null"]
        errorRate:
          type: ["number", "null"]
        activeDays:
          type: number
        firstEventAt:
          type: ["string", "null"]
        lastEventAt:
          type: ["string", "null"]
    ReevaluateOutcome:
      type: object
      required: [agentId, changed, healthScore, verdict, dataSource, rationale]
      properties:
        agentId:
          type: string
        changed:
          type: boolean
        healthScore:
          type: number
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        dataSource:
          type: string
          enum: [telemetry, seeded, mixed, none]
        rationale:
          type: string
        rulesFired:
          type: array
          items:
            type: string